move to `ExPTY.Unix`
